image-service/pyproject.toml
[project]
name = "aigc-image-service"
version = "1.0.0"
description = "基于 Daft + OpenAI 的多模态图像分析服务"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "getdaft>=0.3.0",
    "ray>=2.9.0",
    "openai>=1.0.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "python-multipart>=0.0.6",
    "httpx>=0.26.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
